Return information about the latest dump if ID not provided
